gvsigol_v2.x/gvsigol/requirements.txt
## django
django==1.11.24
djangoajax==2.3.5
httplib2==0.9.2
geopy==1.14.0
gisdata==0.5.4

# Python 2 / 3 compatibility layer
future==0.17.1

## auth
python-ldap==2.5.2
django-auth-ldap==1.2.7

## django rest framework
#djangorestframework
#markdown
#django-filter
#django-rest-auth
#django-formtools
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
django-rest-swagger==2.2.0
django-rest-swagger-swaggerdoc==1.0.post3
django-cors-headers==3.0.2

##Instalamos el driver de PostgreSQL para python:
psycopg2>=2.7,<2.8 --no-binary psycopg2
# *NOTE:* Es necesario instalar previamente la libreria _libpq-dev_
#psycopg2-binary>=2.7,<2.8

##Instalamos Pillow para que django soporte la gestion de images;
Pillow==3.4.2
# *NOTE:* Es necesario instalar previamente la libreria _libjpeg-dev_

## gsconfig
geoserver-restconfig==1.0.6
requests==2.21.0

# async requests
requests-futures==0.9.7
futures==3.2.0 # only needed for Python 2.x

## Scolab tools
spatialiteintrospect==1.2
pygdaltools==1.2.1
# *NOTE:* It requires having gdal tools installed in the system (gdal-bin package)

## Other
xmltodict
six==1.13.0
pytz==2016.3
python-dateutil==2.5.2

numpy==1.12.0

argparse==1.2.1
wsgiref==0.1.2

cmislib==0.5.1
lxml==4.3.0

# Cron para aguas de valencia
schedule==0.4.2

# Ver #3727
openpyxl==2.6.4
# Lectura de DBFs
dbfread==2.0.7
django-excel==0.0.6
pyexcel-xls==0.4.0
pyexcel-xlsx==0.4.0
pyexcel-ods==0.4.0

# Parseo del getCapabilities (baseLayers)
# se instala un pyproj compatible con python 2
pyproj==2.2.2
OWSLib==0.18.0

limesurveyrc2api==1.0.0

suds==0.4

django-jenkins==0.110.0

python-resize-image==1.1.11
django-activity-stream==0.8.0
django-jsonfield==1.2.0
django-jsonfield-compat==0.4.4

# xsendfile wrapper
django-sendfile2==0.4.3

# Comprobar el timestamp de un fichero antes de bajarlo
tzlocal==1.5.1

# language code database
iso-639

# job and queque manager, for download manager and email queues
celery==4.3.0
# harcode indirect celery requirements
kombu==4.5.0
zipp==0.6.0

django-celery-beat==1.6.0

# workaround for indirect dependencies problem when using pip 8
more-itertools<6.0.0

# uwsgi server
uwsgi==2.0.18
gevent==1.4.0
#greenlet==0.4.12

APScheduler==3.6.3

## compatibilidad con python2 julio 2021
# dependencias de drf-yasg
ruamel.ordereddict==0.4.15
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
# dependencias de APScheduler, mock
funcsigs==1.0.2
# dependencias de django-rest-swagger-swaggerdoc, drf-yasg
itypes==1.2.0
